+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Changed
-
+* better LaTeX memoir template
 
 ## [0.0.3] -- 2022-05-27