+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * bugs
 
 ### Changed
-* better LaTeX memoir template
+* nicer LaTeX memoir template
 
 ## [0.0.3] -- 2022-05-27