+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Removed
+
+### Changed
+
+## [0.0.4] -- 2022-06-03
+
+### Added
 * `todo` text command
 * IGT examples as part of manual multipart examples
 * `check` terminal command
@@ -73,7 +81,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[Unreleased]: https://github.com/fmatter/pylingdocs/compare/0.0.3...HEAD
+[Unreleased]: https://github.com/fmatter/pylingdocs/compare/0.0.4...HEAD
+[0.0.4]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.4
 [0.0.3]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.3
 [0.0.2]: https://github.com/fmatter/pylingdocs/releases/tag/0.0.2
 [0.0.1]: https://github.com/fmatter/pylingdocs/releases/tag/v0.0.1